Got rid of dynamic allocation for parse_tle:
This requires the user to provide a pointer to both the orbit structure and the ephemeris structure. This unfortunately required sgp4/sdp4.h to be exported through includes, for the user to do that. Unless an opague pointer is declared in predict.h which wraps the size of the struct, so it can be done.

Also i exported sun.h, because we need the sun coordinated in ECI, and there was no other way to pull that from the primary api